lib/requirements.txt
absl-py==0.7.1
astor==0.7.1
atari-py==0.1.7
aenum==2.1.2
asn1crypto==0.24.0
attrs==18.2.0
baseline==0.2.1
baselines==0.1.5
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.7
dill==0.2.9
entrypoints==0.3
future==0.17.1
gast==0.2.2
glfw==1.7.1
glob2==0.6
google-pasta==0.1.7
grpcio==1.20.0
gym==0.12.1
gym-retro==0.7.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
jeepney==0.4
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
mpi4py==3.0.1
nes-py==6.2.2
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
pbr==5.1.3
Pillow==6.0.0
plotly==4.1.0
progressbar2==3.39.3
psutil==5.6.3
pycparser==2.19
pygame==1.9.6
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
seaborn==0.9.0
SecretStorage==3.1.1
six==1.12.0
soupsieve==1.9.1
stable-baselines==2.7.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.2
Werkzeug==0.15.2
wrapt==1.11.2
zmq==0.0.0
fbchat==1.6.5
keyring==19.0.1
keyrings.alt==3.1.1
# vizdoom==1.1.7
--no-binary=protobuf protobuf

